Fix bootstrap-sass error.
Hey all,

To debug this, I started up the app and saw that the error was pointing
to line 5 in application.scss. I looked at that line and saw that it was
the bootstrap-sass import. I then google the error with "bootstrap sass"
and came across this.

http://stackoverflow.com/questions/22392862/undefined-method-environment-for-nilnilclass-when-importing-bootstrap

It looks like there was a bug in the version of sprockets we are using. So I upgraded it & everything looks good now. :D